TST: Move and rework coarsening test
