+++ src/porepy/applications/test_utils/reference_dense_arrays.py
@@ -2943,3 +2943,615 @@ test_partition["test_cartesian_grids"]["6_5_4"] = {
         ]
     )
 }
+# test_coarsening.py
+test_coarsening: dict[str, dict] = {
+    "test_coarse_grid_3d": {
+        "face_nodes": np.array(
+            [
+                [0, 3, 9, 12],
+                [2, 5, 11, 14],
+                [3, 6, 12, 15],
+                [5, 8, 14, 17],
+                [9, 12, 18, 21],
+                [11, 14, 20, 23],
+                [12, 15, 21, 24],
+                [14, 17, 23, 26],
+                [0, 1, 9, 10],
+                [1, 2, 10, 11],
+                [6, 7, 15, 16],
+                [7, 8, 16, 17],
+                [9, 10, 18, 19],
+                [10, 11, 19, 20],
+                [12, 13, 21, 22],
+                [13, 14, 22, 23],
+                [15, 16, 24, 25],
+                [16, 17, 25, 26],
+                [0, 1, 3, 4],
+                [1, 2, 4, 5],
+                [3, 4, 6, 7],
+                [4, 5, 7, 8],
+                [9, 10, 12, 13],
+                [10, 11, 13, 14],
+                [12, 13, 15, 16],
+                [13, 14, 16, 17],
+                [18, 19, 21, 22],
+                [19, 20, 22, 23],
+                [21, 22, 24, 25],
+                [22, 23, 25, 26],
+            ]
+        )
+    }
+}
+_cell_centers_1 = np.array(
+    [
+        [
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+        ],
+        [
+            1.5,
+            1.5,
+            1.5,
+            1.5,
+            2.5,
+            2.5,
+            2.5,
+            2.5,
+            3.5,
+            3.5,
+            3.5,
+            3.5,
+            4.5,
+            4.5,
+            4.5,
+            4.5,
+        ],
+        [
+            4.5,
+            3.5,
+            2.5,
+            1.5,
+            4.5,
+            3.5,
+            2.5,
+            1.5,
+            4.5,
+            3.5,
+            2.5,
+            1.5,
+            4.5,
+            3.5,
+            2.5,
+            1.5,
+        ],
+    ]
+)
+_cell_centers_2 = np.array(
+    [
+        [
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+        ],
+        [
+            1.5,
+            1.5,
+            1.5,
+            1.5,
+            2.5,
+            2.5,
+            2.5,
+            2.5,
+            3.5,
+            3.5,
+            3.5,
+            3.5,
+            4.5,
+            4.5,
+            4.5,
+            4.5,
+        ],
+        [
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+        ],
+    ]
+)
+_cell_centers_3 = np.array(
+    [
+        [
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+            3.0,
+        ],
+        [
+            1.5,
+            1.5,
+            1.5,
+            1.5,
+            2.5,
+            2.5,
+            2.5,
+            2.5,
+            3.5,
+            3.5,
+            3.5,
+            3.5,
+            4.5,
+            4.5,
+            4.5,
+            4.5,
+        ],
+        [
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+            1.5,
+            2.5,
+            3.5,
+            4.5,
+        ],
+    ]
+)
+_indices_1 = [
+    3,
+    7,
+    11,
+    15,
+    2,
+    6,
+    10,
+    14,
+    1,
+    5,
+    9,
+    13,
+    0,
+    4,
+    8,
+    12,
+    19,
+    23,
+    27,
+    31,
+    18,
+    22,
+    26,
+    30,
+    17,
+    21,
+    25,
+    29,
+    16,
+    20,
+    24,
+    28,
+]
+_faces_1 = [
+    22,
+    25,
+    28,
+    31,
+    40,
+    43,
+    46,
+    49,
+    58,
+    61,
+    64,
+    67,
+    76,
+    79,
+    82,
+    85,
+    288,
+    289,
+    290,
+    291,
+    292,
+    293,
+    294,
+    295,
+    296,
+    297,
+    298,
+    299,
+    300,
+    301,
+    302,
+    303,
+]
+_indices_2 = [
+    0,
+    1,
+    2,
+    3,
+    4,
+    5,
+    6,
+    7,
+    8,
+    9,
+    10,
+    11,
+    12,
+    13,
+    14,
+    15,
+    16,
+    17,
+    18,
+    19,
+    20,
+    21,
+    22,
+    23,
+    24,
+    25,
+    26,
+    27,
+    28,
+    29,
+    30,
+    31,
+]
+_faces_2 = [
+    223,
+    224,
+    225,
+    226,
+    229,
+    230,
+    231,
+    232,
+    235,
+    236,
+    237,
+    238,
+    241,
+    242,
+    243,
+    244,
+    304,
+    305,
+    306,
+    307,
+    308,
+    309,
+    310,
+    311,
+    312,
+    313,
+    314,
+    315,
+    316,
+    317,
+    318,
+    319,
+]
+_indices_3 = [
+    0,
+    4,
+    8,
+    12,
+    1,
+    5,
+    9,
+    13,
+    2,
+    6,
+    10,
+    14,
+    3,
+    7,
+    11,
+    15,
+    16,
+    20,
+    24,
+    28,
+    17,
+    21,
+    25,
+    29,
+    18,
+    22,
+    26,
+    30,
+    19,
+    23,
+    27,
+    31,
+]
+_faces_3 = [
+    22,
+    25,
+    28,
+    31,
+    40,
+    43,
+    46,
+    49,
+    58,
+    61,
+    64,
+    67,
+    76,
+    79,
+    82,
+    85,
+    288,
+    289,
+    290,
+    291,
+    292,
+    293,
+    294,
+    295,
+    296,
+    297,
+    298,
+    299,
+    300,
+    301,
+    302,
+    303,
+]
+test_coarsening["test_coarse_grid_3d_2d_cross"] = {
+    "cell_centers_1": _cell_centers_1,
+    "cell_centers_2": _cell_centers_2,
+    "cell_centers_3": _cell_centers_3,
+    "indices_1": _indices_1,
+    "indices_2": _indices_2,
+    "indices_3": _indices_3,
+    "faces_1": _faces_1,
+    "faces_2": _faces_2,
+    "faces_3": _faces_3,
+}
+_partition = (
+    np.array(
+        [
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            2,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            2,
+            2,
+            3,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            1,
+            2,
+            2,
+            3,
+            3,
+            1,
+            1,
+            1,
+            1,
+            1,
+            2,
+            2,
+            2,
+            3,
+            3,
+            3,
+            1,
+            1,
+            1,
+            1,
+            2,
+            2,
+            2,
+            3,
+            3,
+            3,
+            3,
+            1,
+            1,
+            2,
+            2,
+            2,
+            2,
+            3,
+            3,
+            3,
+            3,
+            3,
+            2,
+            2,
+            2,
+            2,
+            2,
+            3,
+            3,
+            3,
+            3,
+            3,
+            2,
+            2,
+            2,
+            2,
+            2,
+        ]
+    )
+    - 1
+)
+test_coarsening["test_create_partition_2d_cart_cdepth4"] = {
+    "partition": _partition,
+}
+_partition = (
+    np.array(
+        [
+            1,
+            1,
+            1,
+            1,
+            2,
+            4,
+            1,
+            3,
+            2,
+            2,
+            3,
+            3,
+            2,
+            2,
+            3,
+            3,
+            5,
+            4,
+            1,
+            6,
+            4,
+            4,
+            4,
+            3,
+            2,
+            4,
+            7,
+            3,
+            8,
+            8,
+            3,
+            3,
+            5,
+            5,
+            6,
+            6,
+            5,
+            4,
+            7,
+            6,
+            8,
+            7,
+            7,
+            7,
+            8,
+            8,
+            7,
+            9,
+            5,
+            5,
+            6,
+            6,
+            5,
+            5,
+            6,
+            6,
+            8,
+            8,
+            7,
+            9,
+            8,
+            8,
+            9,
+            9,
+        ]
+    )
+    - 1
+)
+test_coarsening["test_create_partition_3d_cart"] = {
+    "partition": _partition,
+}
